Factored DataProtobufConverter out of specs.
First step in removing TensorFlow dependency from the specs
module which is also used by the API service which does not
need to load TensorFlow.
