+++ service/run_tests.py
@@ -90,6 +90,7 @@ _DEFAULT_TEST_MODULES = [
     'learner.brains.action_postprocessor_test',
     'learner.brains.brain_cache_test',
     'learner.brains.continuous_imitation_brain_test',
+    'learner.brains.data_protobuf_converter_test',
     'learner.brains.demonstration_buffer_test',
     'learner.brains.egocentric_test',
     'learner.brains.eval_datastore_test',
